seperate out table-view into a different file
